requirements.txt
flask==2.0.3
werkzeug==2.0.3
google-cloud-pubsub==2.13.0
requests==2.32.2
numpy==1.23.5
pandas==1.5.3
python-dotenv==1.0.0
gunicorn==20.1.0
